requirements.txt
twine==5.1.1
wheel==0.41.2
keyring
keyrings.alt
pylint
git-pylint-commit-hook
numpy
pandas
requests
requests-toolbelt
autopep8
dataclasses
dataclasses-json
pytest
python-dotenv
humanfirst
pytest-cov
ntplib
PyJWT